requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.96.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
pgvector==0.1.8
psycopg2==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
